game history

+++ Backend/src/main/java/com/revature/data/GameHistoryDAO.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface GameHistoryDAO extends GenericDAO<GameHistory> {
 
     public List<GameHistory> getDailyLeaderboard();
-    public List<GameHistory> getByPersonId(Integer id);
+    public List<GameHistory> getByPersonId(Long id);
 
     public GameHistory getByLongId(Long id);
